backend/requirements.txt
# QuantDog dependencies
Flask==3.0.2
python-dotenv==1.0.1
psycopg[binary]==3.2.3
SQLAlchemy==2.0.30
alembic==1.13.1
pydantic==2.10.6
httpx==0.28.1
yfinance==0.2.44
